Starting up localization package
Far from done, and about to change up initial design ideas. The plan is to have a design where we generate assets as we go when we're working on translation, and throw if they're missing in actual production use.

The hard part is that the localization package uses a lot of project-level stuff. Active localization settings are project-level, and the ui populates from this. Conventions for addressable groups are also project-level. Addressable-related stuff also can't be packaged up easily, nor at all in our version if Unity.

On the bright side, for production consumption where we don't modify localization assets, there should be no problem. We just need to be able to read all the assets, and they don't need to be addressable.

For editing, we'll need an interface and ui that allows selecting the project-level localization stuff. This interface will also somehow signal our localization stuff to generate artifacts for the particular localizationsettings, instead of failing for missing stuff.

One minor-ish design detail is returning the project-level stuff back to normal. This would likely require serializing this somewhere. But we'll probably take the easy way out and just allow unsetting everything, allowing users to figure it out themselves. We're the likely only user, so sure this should be fine.