Got initial working implementation of runtime localization generation done
One small issue is that `Enter edit mode` can be a bit flaky, but stopping and entering again works just fine, so don't really care.

Currently, we manually get tables and fail if we can't find them. We also theoretically do it for lookups. This is a problem because throwing in ui generation has been obnoxious in practice, and it would also be nice to use the default behavior of getting a failed string returned from localization. Will try to get that up-and-running next.