Attempting to fix bug where prefab children arent copied correctly
They were being copied as normal objects and not prefab instances. We now currently simply instantiate the object from the original prefab. We don't try to copy over modifications or apply additions or removals, since we already have logic to do that generically. It's also appears to be a pain in the butt to do it thru PrefabUtility. It still remains to be seen if this is sufficient, though.

Also lumped in a similar change to Object.Instantiate in new GameObjects, which is pretty similar to a duplicate operation. It also remains to be seen if this has any point or not.