Add the concept of unavailable version to package manager
It's possible for there to be no available package versions, such as when all versions are prerelease versions, and we're not showing them. We'll still show the package in the list, in that case, but we'll indicate it as unavailable and lock the modification buttons away.

Also upgraded .net just because.