Abstract out common stuff for analyzing playable layers
We're about to write a new analyzer that needs to know about whether or not layers may be disabled, so some refactoring is in order.

Also note that the layer 1/2 analyzer is broken, since transition traversal hasn't been factored out. While we won't be using this in other places yet, the renderer now uses a class instead of a tuple -- a class only provided by this new component. Plus, it'll be nice to put this code there and simplify the analyzer, anyway.