Begin work on editor language switching support
The key feature of this will be installing other languages, normally done through Unity Hub.

Started with the language switching by experimentally using the same UI as in editor preferences, which is not without issues. The main issue is that switching languages itself is kinda borked, which overshadows an NRE that really doesn't hurt anything. This involved creating the interesting element SettingsProviderContainer.

Additionally, added a convention of a common way to clone a VisualTreeAsset. Currently, this just applies a common style.