Move retranslation tracking to CatalogReference, add IRetranslatable
CatalogReference may need a bit of a rebranding, but it kinda naturally belongs there. We want to retranslate if the catalog changes or if the selected locale changes. The initial thought was to put it in LocaleSelector, but that's obviously only going to support just locale changing.

IRetranslatable is a simple interface that may not ever actually be used.