Got a decent part of the extraction code done, plus more T changes
For the T changes, I realized that one/id also always has to be formattable for plurals because it'll usually include the number in it, like every other form (except special zero). This, in turn, caused more overload resolution issues. The big plural T method now no longer has plural forms optional, because it no longer becomes possible to resolve the overload that has differing specialZero types. The idea was that we'd generate the proxy type based on the native language and only exposing the required parameters for that. However, theoretically, we can move specialZero to right after other, then have overloads that include or exclude it.

In order to make extraction smooth, we have some simple attributes. One indicates the method is for localization. These methods then either belong to a property instance, field instance, or are static methods to a class; these targets have an attribute indicating the catalog location and languages supported. Though, we'll likely soon merge them into a single language attribute that points to a po file path. It's then up to the author to basically make sure they're in the same directory to be loaded by EZLocalization.

We rely heavily on Roslyn operation APIs for taking calls to a marked method, and...
* Parsing out the parameters, reading literal values to them (or otherwise failing to extract), processing them based on convention of parameter names.
* Tracing through to find attributes on various method definitions, field declarations, and class definitions.

What's not yet done is actually producing entries from parsed method calls, merging them into an existing file, an writing them out. The proxy type generation also hasn't been started, though the groundwork has been laid.