Begin unit testing GetTextDocument parsing, and fix some bugs
Unit testing far from done, but we've already fixed a few bugs.

As a part of extraction, we need to add some logic to deal with deprecated entries, which further complicates parsing. As such, it's high time we unit tested! Still not sure how we'll implement it yet, but probably won't go the route of parsing the comment, after all. In any case, I'll first see how xgettext or whatever natively does to deprecate entries in order to imitate better. For instance, do comments get re-commented, or just the bits with keywords?