Removed special zero in favor of just having languages use zero
Mainly triggered by other tooling having restrictions on languages' plural forms. The nice thing about it was that languages that don't have zero forms could opt into having a zero form, with overloads to make it ezpz. On the other hand, it's not like there was a scenario to have both zero and special zero going at the same time, so it was a weird-ish design to begin with.

On the bright side, the design is simpler and reduced the number of T overloads.

Also a note this change isn't complete. Hit issues that arise when changing GenerateLanguage plural rules, where we purposely throw. The final bit of this change comes in next commit.