Finish exception audit
Was expecting to more gracefully handle exceptions, but we're pretty okay. Well, sorta.
For extraction, exceptions just arent a big deal. They show up in the logs, can get fixed, and do wanna add a button to rerun.
For catalog parsing, the only issue is that we get exceptions in cctors, which unity doesn't like that much. Otherwise, all of the exceptions we generate are where we have no reasonable way of providing some sort of fail-safe, so it's fine.
And for T, we no longer generate exceptions there.