Fix integration tests for plurals
Added an OverwriteEntry operation to the doc builder, which will take the values from a built entry and overwrite what's in an existing entry.
It also throws if the existing entry isn't found. In addition to the test convenience of the throw, it's a generally nice way to keep comments and whatnot. Addiitionally, the merge variant was insufficient because an existing entry means the existing values are used, while we were trying to update the values.

Was found while doing other refactorings.