Break CatalogReference's dependency on CatalogLocaleSynchronizer
This was done by making it CatalogLocaleSynchronizer's responsibility to tell CatalogReference what locale it has. Furthermore, we now have two key locales: current locale and intended locale. Intended locale is the old equivalent CatalogLocaleSynchronizer.SelectedLocale. Current locale is what the GetTextCatalog was able to use at the time. If a reload happens, we prefer the intended locale, followed by the current locale, followed by native.

As a consequence...
* EZLocalization now calls CatalogLocaleSynchronizer's locale selection methods, and CatalogReference's are now removed. Note that one is still left because of other simultaneous changes coming in next commit.
* LocaleSelectionUI now has a static ReportChange called by CatalogReference. Not ideal, but not worth designing a better solution yet.