Fix localization bug where non-native pref doesnt stick after reload
We work around the actual bug by changing initialization order of catalogreference and synchronizer. In the past, these had specific orders, but that went away, just no reason to change it. (forgot to update a comment tho)

The actual bug is a hard fix and low-pri. As long as every catalog in the synchronizer key has the same set of locales, we'll have no problems.